Migrate container.
