+++ crates/macros/src/common/mod.rs
@@ -3,6 +3,7 @@ mod field;
 mod field_value;
 mod variant;
 
+pub use container::*;
 pub use field::*;
 pub use field_value::*;
 pub use variant::*;
